requirements.txt
cycler
future
lap
matplotlib
motmetrics
numpy
opencv-python
pandas
sacred
scikit-image
scipy
seaborn
submitit
tqdm
visdom
pillow

--extra-index-url https://download.pytorch.org/whl/cu121
torch
torchvision
torchaudio